TECH-349: Typography type organization
