+++ src/components/Typography/Typography.tsx
@@ -1,55 +1,7 @@
-import React, {FunctionComponent} from 'react';
+import React from 'react';
 import clsx from 'clsx';
 import './Typography.scss';
-
-export type TypographyVariant = 'title' | 'heading' | 'subheading' | 'body' | 'caption' | 'button';
-export const variants = ['title', 'heading', 'subheading', 'body', 'caption', 'button'];
-
-export type TypographyWeight = 'default' | 'bold' | 'semiBold' | 'light';
-export const weights = ['default', 'bold', 'semiBold', 'light'];
-
-interface TypographyProps {
-    /**
-     * Content of the component
-     */
-    children?: React.ReactNode;
-    /**
-     * Custom classname to use
-     */
-    className?: string;
-    /**
-     * The component used for the root node
-     */
-    component?: string;
-    /**
-     * Variant to use
-     */
-    variant?: TypographyVariant;
-    /**
-     * The weight of the font to use
-     */
-    weight?: TypographyWeight;
-    /**
-     * Should the text be displayed in italic
-     */
-    isItalic?: boolean;
-    /**
-     * Should the text be displayed in upper case
-     */
-    isUpperCase?: boolean;
-    /**
-     * Should the text be displayed with a line-through
-     */
-    hasLineThrough?: boolean;
-    /**
-     * Does the children contain HTML markup
-     */
-    isHtml?: boolean;
-    /**
-     * No wrapping for text
-     */
-    isNowrap?: boolean;
-}
+import {TypographyProps} from './Typography.types';
 
 // IsHtml prop should eventually be removed (along with this function) as children supports all React Node types
 // including strings and HTML markup
@@ -62,7 +14,7 @@ const filterOutIsHtml = (props: TypographyProps) => {
     return newProps;
 };
 
-export const Typography: FunctionComponent<TypographyProps> = ({
+export const Typography: React.FC<TypographyProps> = ({
     children = '',
     component = 'p',
     variant = 'body',
@@ -74,7 +26,7 @@ export const Typography: FunctionComponent<TypographyProps> = ({
     isNowrap = false,
     isHtml = false,
     ...props
-}: TypographyProps) => {
+}) => {
     if (!children) {
         return null
     }
