+++ src/components/Typography/Typography.types.ts
@@ -1,0 +1,50 @@
+import React from 'react';
+
+type TypographyVariant = 'title' | 'heading' | 'subheading' | 'body' | 'caption' | 'button';
+export const variants = ['title', 'heading', 'subheading', 'body', 'caption', 'button'];
+
+type TypographyWeight = 'default' | 'bold' | 'semiBold' | 'light';
+export const weights = ['default', 'bold', 'semiBold', 'light'];
+
+export type TypographyProps = {
+    /**
+     * Content of the component
+     */
+    children?: React.ReactNode;
+    /**
+     * Custom classname to use
+     */
+    className?: string;
+    /**
+     * The component used for the root node
+     */
+    component?: string;
+    /**
+     * Variant to use
+     */
+    variant?: TypographyVariant;
+    /**
+     * The weight of the font to use
+     */
+    weight?: TypographyWeight;
+    /**
+     * Should the text be displayed in italic
+     */
+    isItalic?: boolean;
+    /**
+     * Should the text be displayed in upper case
+     */
+    isUpperCase?: boolean;
+    /**
+     * Should the text be displayed with a line-through
+     */
+    hasLineThrough?: boolean;
+    /**
+     * Does the children contain HTML markup
+     */
+    isHtml?: boolean;
+    /**
+     * No wrapping for text
+     */
+    isNowrap?: boolean;
+}
